requirements.txt
alabaster @ file:///home/ktietz/src/ci/alabaster_1611921544520/work
altair==4.1.0
asn1crypto @ file:///tmp/build/80754af9/asn1crypto_1596577642040/work
astroid @ file:///C:/ci/astroid_1628063237844/work
attrs @ file:///opt/conda/conda-bld/attrs_1642510447205/work
Babel @ file:///tmp/build/80754af9/babel_1607110387436/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
backports.zoneinfo==0.2.1
bleach @ file:///tmp/build/80754af9/bleach_1626470256873/work
blinker==1.4
brotlipy==0.7.0
cachetools==4.2.4
certifi==2021.5.30
cffi @ file:///C:/ci/cffi_1625831763871/work
chardet @ file:///C:/ci/chardet_1607706912142/work
click==8.0.4
cloudpickle @ file:///Users/ktietz/demo/mc3/conda-bld/cloudpickle_1629142150447/work
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
cryptography==2.3.1
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
docopt==0.6.2
docutils @ file:///C:/ci/docutils_1620828217025/work
entrypoints==0.3
gitdb==4.0.9
GitPython==3.1.18
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
imagesize @ file:///tmp/build/80754af9/imagesize_1637939814114/work
importlib-metadata @ file:///C:/ci/importlib-metadata_1631916885413/work
importlib-resources==5.4.0
ipykernel @ file:///C:/ci/ipykernel_1596206775157/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1593446240034/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets==7.7.0
isort @ file:///tmp/build/80754af9/isort_1628603791788/work
jedi==0.17.0
Jinja2 @ file:///tmp/build/80754af9/jinja2_1612213139570/work
joblib @ file:///tmp/build/80754af9/joblib_1613502643832/work
json5 @ file:///tmp/build/80754af9/json5_1624432770122/work
jsonschema @ file:///Users/ktietz/demo/mc3/conda-bld/jsonschema_1630511932244/work
jupyter-client==5.3.3
jupyter-core==4.5.0
jupyterlab==1.2.6
jupyterlab-server==1.0.0
jupyterlab-widgets==1.1.0
kiwisolver @ file:///C:/ci/kiwisolver_1612282446297/work
lazy-object-proxy @ file:///C:/ci/lazy-object-proxy_1616529300868/work
MarkupSafe @ file:///C:/ci/markupsafe_1621528313524/work
matplotlib @ file:///C:/ci/matplotlib-suite_1613408055530/work
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
nbconvert==5.5.0
nbformat @ file:///tmp/build/80754af9/nbformat_1617383369282/work
notebook==5.3.1
numpy==1.19.5
numpydoc @ file:///tmp/build/80754af9/numpydoc_1605117425582/work
olefile==0.46
packaging @ file:///tmp/build/80754af9/packaging_1611952188834/work
pandas==1.1.5
pandocfilters @ file:///opt/conda/conda-bld/pandocfilters_1643405455980/work
parso @ file:///tmp/build/80754af9/parso_1596826841367/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==8.4.0
pipreqs==0.4.11
plotly @ file:///opt/conda/conda-bld/plotly_1646671701182/work
prometheus-client @ file:///tmp/build/80754af9/prometheus_client_1623189609245/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
protobuf==3.19.4
psutil @ file:///C:/ci/psutil_1612298125479/work
pyarrow==6.0.1
pycodestyle @ file:///tmp/build/80754af9/pycodestyle_1615748559966/work
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pydeck==0.6.2
pyflakes @ file:///tmp/build/80754af9/pyflakes_1636644436481/work
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
pylint @ file:///C:/ci/pylint_1627536908981/work
Pympler==1.0.1
pyOpenSSL==19.0.0
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
pyrsistent @ file:///C:/ci/pyrsistent_1600141799440/work
PySocks @ file:///C:/ci/pysocks_1605305839978/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32==303
pywinpty==1.1.6
pyzmq @ file:///C:/ci/pyzmq_1628276569449/work
QtAwesome @ file:///tmp/build/80754af9/qtawesome_1602272867890/work
qtconsole @ file:///tmp/build/80754af9/qtconsole_1600870028330/work
QtPy @ file:///opt/conda/conda-bld/qtpy_1643087291789/work
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
rope @ file:///opt/conda/conda-bld/rope_1643788605236/work
scikit-learn @ file:///C:/ci/scikit-learn_1622739439730/work
scipy @ file:///C:/ci/scipy_1597675683670/work
semver==2.13.0
Send2Trash @ file:///tmp/build/80754af9/send2trash_1632406701022/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
snowballstemmer @ file:///tmp/build/80754af9/snowballstemmer_1611258885636/work
Sphinx @ file:///tmp/build/80754af9/sphinx_1616268783226/work
sphinxcontrib-applehelp @ file:///home/ktietz/src/ci/sphinxcontrib-applehelp_1611920841464/work
sphinxcontrib-devhelp @ file:///home/ktietz/src/ci/sphinxcontrib-devhelp_1611920923094/work
sphinxcontrib-htmlhelp @ file:///home/ktietz/src/ci/sphinxcontrib-htmlhelp_1611920974801/work
sphinxcontrib-jsmath @ file:///home/ktietz/src/ci/sphinxcontrib-jsmath_1611920942228/work
sphinxcontrib-qthelp @ file:///home/ktietz/src/ci/sphinxcontrib-qthelp_1611921055322/work
sphinxcontrib-serializinghtml @ file:///tmp/build/80754af9/sphinxcontrib-serializinghtml_1624451540180/work
spyder==3.2.8
streamlit==1.8.1
tenacity @ file:///C:/ci/tenacity_1632759879010/work
terminado==0.12.1
testpath @ file:///home/ktietz/src/ci/testpath_1611930608132/work
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
toolz==0.11.2
tornado==5.1.1
traitlets @ file:///C:/ci/traitlets_1632759765830/work
typed-ast @ file:///C:/ci/typed-ast_1624953776872/work
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
tzdata==2022.1
tzlocal==4.2
urllib3 @ file:///opt/conda/conda-bld/urllib3_1643638302206/work
validators==0.18.2
watchdog==2.1.7
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
widgetsnbextension==3.6.0
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306197271/work
wincertstore==0.2
wrapt==1.12.1
yarg==0.1.9
zipp @ file:///tmp/build/80754af9/zipp_1633618647012/work
